+++ packages/zowe-explorer-api/src/extend/MainframeInteraction.ts
@@ -501,6 +501,16 @@ export namespace MainframeInteraction {
         issueTsoCommandWithParms?(command: string, parms?: zostso.IStartTsoParms): Promise<zostso.IIssueResponse>;
 
         /**
+         * Issues a TSO Command without the need for account information
+         *
+         * @param {string} command
+         * @param {zostso.IStartTsoParms} parms
+         * @returns {Promise<zostso.IIssueResponse>}
+         * @memberof ICommand
+         */
+        issueTsoCmdWithParms?(command: string, parms?: zostso.IStartTsoParms): Promise<zostso.IIssueResponse>;
+
+        /**
          * Issues a MVS Command and returns a Console Command API response.
          *
          * @param {string} command