+++ packages/zowe-explorer/src/configuration/Constants.ts
@@ -70,6 +70,7 @@ export class Constants {
     public static readonly SETTINGS_DS_DEFAULT_EXTENDED = "zowe.ds.default.extended";
     public static readonly SETTINGS_DS_DEFAULT_PS = "zowe.ds.default.ps";
     public static readonly SETTINGS_DS_TEMPLATES = "zowe.ds.templates";
+    public static readonly SETTINGS_COMMANDS_INTEGRATED_TERMINALS = "zowe.commands.useIntegratedTerminals";
     public static readonly SETTINGS_COMMANDS_ALWAYS_EDIT = "zowe.commands.alwaysEdit";
     public static readonly SETTINGS_AUTOMATIC_PROFILE_VALIDATION = "zowe.automaticProfileValidation";
     public static readonly SETTINGS_SECURE_CREDENTIALS_ENABLED = "zowe.security.secureCredentialsEnabled";
@@ -95,7 +96,9 @@ export class Constants {
         "Zowe-Default-Datasets-PDS": Constants.SETTINGS_DS_DEFAULT_PDS,
         "Zowe-Default-Datasets-Extended": Constants.SETTINGS_DS_DEFAULT_EXTENDED,
         "Zowe-Default-Datasets-PS": Constants.SETTINGS_DS_DEFAULT_PS,
-        "Zowe Commands: History": PersistenceSchemaEnum.Commands,
+        "Zowe MVS Commands: History": PersistenceSchemaEnum.MvsCommands,
+        "Zowe TSO Commands: History": PersistenceSchemaEnum.TsoCommands,
+        "Zowe Unix Commands: History": PersistenceSchemaEnum.UssCommands,
         "Zowe Commands: Always edit": Constants.SETTINGS_COMMANDS_ALWAYS_EDIT,
         "Zowe-Automatic-Validation": Constants.SETTINGS_AUTOMATIC_PROFILE_VALIDATION,
         "Zowe-DS-Persistent": PersistenceSchemaEnum.Dataset,
@@ -263,5 +266,6 @@ export class Constants {
         DS: "ds-panel-tab",
         USS: "uss-panel-tab",
         JOBS: "jobs-panel-tab",
+        CMDS: "cmds-panel-tab",
     };
 }